r10k support intial commit
Prior to this commit I had not really tested r10k
This commit gets thinsg setup in such a way to allow for dynamic environments
and a jiminy folder in the confdir. I thought about modules but as I am force=>true on vcsrepo
I did not want to wipe people's modules out accidently.